refactor testing directories add test for body extractor
