+++ src/test/java/de/usd/cstchef/operations/arithmetic/DivideListTest.java
@@ -1,4 +1,4 @@
-package de.usd.cstchef.operations;
+package de.usd.cstchef.operations.arithmetic;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -7,12 +7,13 @@ import burp.CstcObjectFactory;
 import burp.api.montoya.core.ByteArray;
 import de.usd.cstchef.Delimiter;
 import de.usd.cstchef.operations.Operation.OperationInfos;
+import de.usd.cstchef.operations.Operation;
 import de.usd.cstchef.operations.OperationCategory;
-import de.usd.cstchef.operations.arithmetic.Median;
-import de.usd.cstchef.utils.UnitTestObjectFactory;
+import de.usd.cstchef.operations.arithmetic.DivideList;
+import de.usd.cstchef.testutils.UnitTestObjectFactory;
 
 @OperationInfos(name = "Test", category = OperationCategory.ARITHMETIC, description = "Test class")
-public class MedianTest extends Median
+public class DivideListTest extends DivideList
 {
     private String delimiter;
     private boolean isFloat;
@@ -33,27 +34,51 @@ public class MedianTest extends Median
     }
 
     @Test
-    public void CommaMedianTest() throws Exception
+    public void CommaDivideTest() throws Exception
     {
         delimiter = "Comma";
         isFloat = false;
 
-        String testValue = "1,2,3,4,5";
+        String testValue = "8,2,4";
         ByteArray result = perform(factory.createByteArray(testValue), null);
 
-        assert result.toString().equals("3");
+        assert result.toString().equals("1");
     }
 
     @Test
-    public void CommaMedianFloatTest() throws Exception
+    public void CommaDivideFloatTest() throws Exception
     {
         delimiter = "Comma";
         isFloat = true;
 
-        String testValue = "1,2,3.5,4,5";
+        String testValue = "8,2,4,2";
         ByteArray result = perform(factory.createByteArray(testValue), null);
 
-        assert result.toString().equals("3.5");
+        assert result.toString().equals("0.5");
+    }
+
+    @Test
+    public void SpaceDivideTest() throws Exception
+    {
+        delimiter = "Space";
+        isFloat = false;
+
+        String testValue = "8 2 4 0.5";
+        ByteArray result = perform(factory.createByteArray(testValue), null);
+
+        assert result.toString().equals("2");
+    }
+
+    @Test
+    public void SpaceDivideFloatTest() throws Exception
+    {
+        delimiter = "Space";
+        isFloat = true;
+
+        String testValue = "8 2 4 4 0.5";
+        ByteArray result = perform(factory.createByteArray(testValue), null);
+
+        assert result.toString().equals("0.5");
     }
 
     @Before
