+++ src/test/java/de/usd/cstchef/operations/arithmetic/MeanTest.java
@@ -1,4 +1,4 @@
-package de.usd.cstchef.operations;
+package de.usd.cstchef.operations.arithmetic;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -7,12 +7,13 @@ import burp.CstcObjectFactory;
 import burp.api.montoya.core.ByteArray;
 import de.usd.cstchef.Delimiter;
 import de.usd.cstchef.operations.Operation.OperationInfos;
+import de.usd.cstchef.operations.Operation;
 import de.usd.cstchef.operations.OperationCategory;
-import de.usd.cstchef.operations.arithmetic.Sum;
-import de.usd.cstchef.utils.UnitTestObjectFactory;
+import de.usd.cstchef.operations.arithmetic.Mean;
+import de.usd.cstchef.testutils.UnitTestObjectFactory;
 
 @OperationInfos(name = "Test", category = OperationCategory.ARITHMETIC, description = "Test class")
-public class SumTest extends Sum
+public class MeanTest extends Mean
 {
     private String delimiter;
     private boolean isFloat;
@@ -33,51 +34,51 @@ public class SumTest extends Sum
     }
 
     @Test
-    public void CommaSumTest() throws Exception
+    public void CommaMeanTest() throws Exception
     {
         delimiter = "Comma";
         isFloat = false;
 
-        String testValue = "1,2,3,4,5,6";
+        String testValue = "8,2,5";
         ByteArray result = perform(factory.createByteArray(testValue), null);
 
-        assert result.toString().equals("21");
+        assert result.toString().equals("5");
     }
 
     @Test
-    public void CommaSumFloatTest() throws Exception
+    public void CommaMeanFloatTest() throws Exception
     {
         delimiter = "Comma";
         isFloat = true;
 
-        String testValue = "1,2,3,4,5,6";
+        String testValue = "0.2,0.3,0.4,0.1";
         ByteArray result = perform(factory.createByteArray(testValue), null);
 
-        assert result.toString().equals("21.0");
+        assert result.toString().equals("0.25");
     }
 
     @Test
-    public void SpacesumTest() throws Exception
+    public void SpaceMeanTest() throws Exception
     {
         delimiter = "Space";
         isFloat = false;
 
-        String testValue = "1.0 2.1 3.2 4.3 5.4 6.5";
+        String testValue = "8 2 5";
         ByteArray result = perform(factory.createByteArray(testValue), null);
 
-        assert result.toString().equals("23");
+        assert result.toString().equals("5");
     }
 
     @Test
-    public void SpacesumFloatTest() throws Exception
+    public void SpaceDivideFloatTest() throws Exception
     {
         delimiter = "Space";
         isFloat = true;
 
-        String testValue = "1.0 2.1 3.2 4.3 5.4 6.5";
+        String testValue = "0.2 0.3 0.4 0.1";
         ByteArray result = perform(factory.createByteArray(testValue), null);
 
-        assert result.toString().equals("22.5");
+        assert result.toString().equals("0.25");
     }
 
     @Before
