+++ src/test/java/de/usd/cstchef/operations/arithmetic/MedianTest.java
@@ -1,4 +1,4 @@
-package de.usd.cstchef.operations;
+package de.usd.cstchef.operations.arithmetic;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -7,12 +7,13 @@ import burp.CstcObjectFactory;
 import burp.api.montoya.core.ByteArray;
 import de.usd.cstchef.Delimiter;
 import de.usd.cstchef.operations.Operation.OperationInfos;
+import de.usd.cstchef.operations.Operation;
 import de.usd.cstchef.operations.OperationCategory;
-import de.usd.cstchef.operations.arithmetic.MultiplyList;
-import de.usd.cstchef.utils.UnitTestObjectFactory;
+import de.usd.cstchef.operations.arithmetic.Median;
+import de.usd.cstchef.testutils.UnitTestObjectFactory;
 
 @OperationInfos(name = "Test", category = OperationCategory.ARITHMETIC, description = "Test class")
-public class MultiplyListTest extends MultiplyList
+public class MedianTest extends Median
 {
     private String delimiter;
     private boolean isFloat;
@@ -33,52 +34,27 @@ public class MultiplyListTest extends MultiplyList
     }
 
     @Test
-    public void CommaMultiplyTest() throws Exception
+    public void CommaMedianTest() throws Exception
     {
         delimiter = "Comma";
         isFloat = false;
 
-        String testValue = "1,2,3,4,5,6";
+        String testValue = "1,2,3,4,5";
         ByteArray result = perform(factory.createByteArray(testValue), null);
 
-        assert result.toString().equals("720");
+        assert result.toString().equals("3");
     }
 
     @Test
-    public void CommaMultiplyFloatTest() throws Exception
+    public void CommaMedianFloatTest() throws Exception
     {
         delimiter = "Comma";
         isFloat = true;
 
-        String testValue = "3,0.5,0.5";
+        String testValue = "1,2,3.5,4,5";
         ByteArray result = perform(factory.createByteArray(testValue), null);
 
-        assert result.toString().equals("0.75");
-    }
-
-    @Test
-    public void SpaceMultiplyTest() throws Exception
-    {
-        delimiter = "Space";
-        isFloat = false;
-
-        String testValue = "1 2 3 4 5 6";
-        ByteArray result = perform(factory.createByteArray(testValue), null);
-
-        System.out.println(result.toString());
-        assert result.toString().equals("720");
-    }
-
-    @Test
-    public void SpaceMultiplyFloatTest() throws Exception
-    {
-        delimiter = "Space";
-        isFloat = true;
-
-        String testValue = "3 0.5 0.5";
-        ByteArray result = perform(factory.createByteArray(testValue), null);
-
-        assert result.toString().equals("0.75");
+        assert result.toString().equals("3.5");
     }
 
     @Before
