+++ src/test/java/de/usd/cstchef/operations/arithmetic/MultiplyListTest.java
@@ -1,4 +1,4 @@
-package de.usd.cstchef.operations;
+package de.usd.cstchef.operations.arithmetic;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -7,12 +7,13 @@ import burp.CstcObjectFactory;
 import burp.api.montoya.core.ByteArray;
 import de.usd.cstchef.Delimiter;
 import de.usd.cstchef.operations.Operation.OperationInfos;
+import de.usd.cstchef.operations.Operation;
 import de.usd.cstchef.operations.OperationCategory;
-import de.usd.cstchef.operations.arithmetic.DivideList;
-import de.usd.cstchef.utils.UnitTestObjectFactory;
+import de.usd.cstchef.operations.arithmetic.MultiplyList;
+import de.usd.cstchef.testutils.UnitTestObjectFactory;
 
 @OperationInfos(name = "Test", category = OperationCategory.ARITHMETIC, description = "Test class")
-public class DivideListTest extends DivideList
+public class MultiplyListTest extends MultiplyList
 {
     private String delimiter;
     private boolean isFloat;
@@ -33,51 +34,52 @@ public class DivideListTest extends DivideList
     }
 
     @Test
-    public void CommaDivideTest() throws Exception
+    public void CommaMultiplyTest() throws Exception
     {
         delimiter = "Comma";
         isFloat = false;
 
-        String testValue = "8,2,4";
+        String testValue = "1,2,3,4,5,6";
         ByteArray result = perform(factory.createByteArray(testValue), null);
 
-        assert result.toString().equals("1");
+        assert result.toString().equals("720");
     }
 
     @Test
-    public void CommaDivideFloatTest() throws Exception
+    public void CommaMultiplyFloatTest() throws Exception
     {
         delimiter = "Comma";
         isFloat = true;
 
-        String testValue = "8,2,4,2";
+        String testValue = "3,0.5,0.5";
         ByteArray result = perform(factory.createByteArray(testValue), null);
 
-        assert result.toString().equals("0.5");
+        assert result.toString().equals("0.75");
     }
 
     @Test
-    public void SpaceDivideTest() throws Exception
+    public void SpaceMultiplyTest() throws Exception
     {
         delimiter = "Space";
         isFloat = false;
 
-        String testValue = "8 2 4 0.5";
+        String testValue = "1 2 3 4 5 6";
         ByteArray result = perform(factory.createByteArray(testValue), null);
 
-        assert result.toString().equals("2");
+        System.out.println(result.toString());
+        assert result.toString().equals("720");
     }
 
     @Test
-    public void SpaceDivideFloatTest() throws Exception
+    public void SpaceMultiplyFloatTest() throws Exception
     {
         delimiter = "Space";
         isFloat = true;
 
-        String testValue = "8 2 4 4 0.5";
+        String testValue = "3 0.5 0.5";
         ByteArray result = perform(factory.createByteArray(testValue), null);
 
-        assert result.toString().equals("0.5");
+        assert result.toString().equals("0.75");
     }
 
     @Before
