+++ src/test/java/de/usd/cstchef/testutils/UnitTestObjectFactory.java
@@ -1,4 +1,4 @@
-package de.usd.cstchef.utils;
+package de.usd.cstchef.testutils;
 
 import burp.CstcObjectFactory;
 import burp.api.montoya.core.ByteArray;
